Added support for specifying implementation types for a mapped collection.
In most simple cases, the generic type of an item in a collection is the same as the entity type that is mapped. However, in some cases, it is necessary that the generic type is different, i.e. a super type, of the actual implementation type. Until now, implementation type and target type of a mapped collection where implicltly set to the same type. Now, a new attribute on the @Collection annotation allows to specify an optional implementation type, enabling constructs like this:

@Collection(targetType=MyInterface.class, implementationType=MyClass.class)
List<MyInterface> list;

This assumes that MyClass is a subtype of MyInterface.